+++ src/groups/special_euclidean_group.jl
@@ -183,6 +183,15 @@ function ManifoldsBase.tangent_vector_type(
     return SpecialEuclideanProductTangentVector{T}
 end
 
+function point_type(::SpecialEuclideanGroup, ::Type{SpecialEuclideanProductTangentVector})
+    return SpecialEuclideanProductPoint
+end
+function point_type(
+    ::SpecialEuclideanGroup, ::Type{SpecialEuclideanProductTangentVector{T}}
+) where {T}
+    return SpecialEuclideanProductPoint{T}
+end
+
 # This union we can also use for the matrix case where we do not care
 
 function SpecialEuclideanGroup(n::Int; variant::Symbol=:left, kwargs...)
@@ -558,6 +567,27 @@ end
 
 # default: Do nothing
 init_constants!(::AbstractManifold, gX) = gX
+
+#overwrite default inner since here the access is a bit tricky.
+function ManifoldsBase.inner(
+    𝔤::LieAlgebra{ℝ,<:SpecialEuclideanGroupOperation,<:SpecialEuclideanGroup},
+    X::AbstractMatrix,
+    Y::AbstractMatrix,
+)
+    G = base_lie_group(𝔤)
+    SOn, Tn = _SOn_and_Tn(G)
+    i1 = inner(
+        LieAlgebra(SOn),
+        submanifold_component(𝔤, X, :Rotation),
+        submanifold_component(𝔤, Y, :Rotation),
+    )
+    i2 = inner(
+        LieAlgebra(Tn),
+        submanifold_component(𝔤, X, :Translation),
+        submanifold_component(𝔤, Y, :Translation),
+    )
+    return i1 + i2
+end
 
 _doc_inv_SEn = """
     inv(G::SpecialEuclideanGroup, g)